pyproject.toml
[tool.poetry]
name = "web-scrapper"
version = "0.1.0"
description = ""
authors = ["James Nornubari <nornubariconfidence@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
openai = "^1.35.13"
sqlalchemy = "^2.0.31"
hydra-core = "^1.3.2"
omegaconf = "^2.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
